api/requirements.txt
Flask
matplotlib
https://download.pytorch.org/whl/cpu/torch-1.0.1.post2-cp36-cp36m-linux_x86_64.whl
torchvision
numpy
Pillow
Werkzeug
uwsgi